pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crew_ia"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "crewai>=0.51",
  "pydantic>=2",
  "pydantic-settings>=2",
  "networkx>=3",
  "httpx>=0.27",
  "anyio>=4",
  "uvicorn>=0.30",
  "fastapi>=0.111",
  "rich>=13",
  "python-dotenv>=1",
  "opentelemetry-sdk>=1.26",
  "opentelemetry-exporter-otlp>=1.26",
  "pytest>=8",
  "aiofiles>=24.1.0"
]

[tool.setuptools]
# le code vit à la racine (pas de src/)
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "core*"]
exclude = ["tests*", "examples*", "api*", "dashboard*", ".*"]

[tool.pytest.ini_options]
addopts = "-q"
